Rename constant for empty context
This name change is only cosmetic, but am updating after it caused some
confusion. The flag_name always needs to be provided, but the context
does not.

We've been using 'undefined_key' for the empty context. The comment and
name are misleading, because 'treatment key' suggests it's the name of
the flag. Let's be clear this stands in for the context.