Add configuration option for base_data_lambda
It's common for projects to want to toggle flags based a handful of
attributes. For Flag A, we'll use attribute foo and for Flag B, we'll
use attribute bar.

But when happens when we actually want to use attribute foo for Flag B.
Right now, this requires a code change. What if instead we can just pass
a record to ThisFeature and the project can specifiy all the common
attributes we can about given that record type?

This PR makes that change. For now, it only works with Split.io

TODO: update README.