`get_treatment` and `get_treatment_with_config` interface for Split and Memory Adapters (#83)
# Context 

[INS-1225](https://hoverinc.atlassian.net/browse/INS-1225)


As the Split Adapter is written currently, it's only capable of reading
from Split flags the value of being `on`, `off`, or in the `control`
group. But Split features can have _many_ options for treatments. We
need to update the Adapters to be able to just retrieve the treatment
value, if we want a fuller use of SplitIO's capabilities.

# Changes 

- expanding the `ThisFeature::Adapters::SplitIo` interface by adding
`#treatment_value` and `#treatment_config` methods, which uses SplitIO's
`get_treatment_with_config` method and returns its `config` or treatment
name as `value`
- `ThisFeature::Flag#treatment_value` method that wraps the adapter's
calling of the same method
- `#treatment_value` equivalent in `ThisFeature::Adapters::Memory` so
this can be used in all environments
- for treatment values to be configured, there is also a
`#enable_treatment!()` method, which is similar to `#on!` which adds to
the local storage of flag configurations and their required contexts
(for which User id, etc)
- I've purposely omitted
`ThisFeature::Adapters::Memory#treatment_config`, because I wanted some
feedback on the `Memory#treatment_value` implementation first
- updates to README for Memory
- if my understanding is correct, this should be a non-breaking change,
because it only include additions to the gem
- TODO: update version, when changes finalized



[INS-1225]:
https://hoverinc.atlassian.net/browse/INS-1225?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Heidi Tsang <tsa.heidi@gmail.com>
Co-authored-by: Heidi Tsang <heidi.tsang@hover.to>